Fix: Add password validation
