+++ lesson26/src/index.html
@@ -5,7 +5,7 @@
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
         <link rel="stylesheet" href="./css/reset.css">
         <link rel="stylesheet" href="./css/style.css">
-        <script src="./js/login.js" defer></script>
+        <script type="module" src="./js/login.js"></script>
         <title>Login</title>
     </head>
     <body>
