+++ lesson26/src/js/login.js
@@ -1,34 +1,46 @@
+import { validationOptions } from "./validation";
+
 const userIdOfInput = document.querySelector(".js-form-userid");
 const passwordOfInput = document.querySelector(".js-form-password");
 const formElements = [userIdOfInput,passwordOfInput];
 const submitButton = document.getElementById("js-submit-button");
 
+const addInvalidClass = target => target.classList.add("invalid");
+const removeInvalidClass = target => target.classList.remove("invalid");
+const showErrorMessage = target => target.nextElementSibling.textContent = validationOptions[target.id].errorMessage;
 const isNotEmptyOfInput = target => target.value.trim() !== "";
-const isNotEmptyOfAllInput = () => {
-    let result = false;
-    formElements.forEach(element => {
-        element.value.trim() !== "" ? result = true : result = false;
-    })
-    return result;
+const isValidFormInput = target => validationOptions[target.id].isValid();
+
+const checkFormValidityToEnableSubmitButton = () => {
+    const invalidItem = document.getElementsByClassName("invalid");
+    submitButton.disabled = invalidItem.length === 0 ? false : true;
 }
 
-const checkFormValidityToEnableSubmitButton = () => submitButton.disabled = isNotEmptyOfAllInput() ? false : true;
-
 formElements.forEach(element => {
+    element.classList.add("invalid");
+
     element.addEventListener("blur", (e) => {
         checkFormValidityInBlur(e.target);
     });
 });
 
-const checkFormValidityInBlur = target => {
+const checkFormValidityInBlur = (target) => {
     submitButton.disabled = true;
 
     if (!isNotEmptyOfInput(target)) {
+        addInvalidClass(target);
         target.nextElementSibling.textContent = "入力してください";
         return;
     }
 
+    if (target.id === "password" && !isValidFormInput(target)) {
+        showErrorMessage(target);
+        addInvalidClass(target);
+        return;
+    }
+
     target.nextElementSibling.textContent = "";
+    removeInvalidClass(target);
     checkFormValidityToEnableSubmitButton();
 };
 
@@ -59,7 +71,7 @@ const checkToRegistered = (inputData) => {
             password : "N302aoe3"
         }
 
-        if (inputData.userId === registeredData.name || inputData.userId === registeredData.email && inputData.password === registeredData.password) {
+        if ((inputData.userId === registeredData.name || inputData.userId === registeredData.email) && (inputData.password === registeredData.password)) {
             resolve({ token: "fafae92rfjafa03", ok: true, code: 200 });
         } else {
             reject({ ok: false, code: 401 });
