+++ lesson26/src/js/register.js
@@ -1,3 +1,5 @@
+import { validationOptions } from "./validation";
+
 const bodyElement = document.querySelector("body");
 const checkboxLink = document.getElementById("js-checkbox-link");
 const modalCloseButton = document.getElementById("js-modal-close-button");
@@ -49,41 +51,10 @@ const setCheckedAttributeToCheckbox = ([entry]) => {
 const observer = new IntersectionObserver(setCheckedAttributeToCheckbox, observerOptions);
 observer.observe(observerTarget);
 
-
 const nameOfInput = document.querySelector(".js-form-name");
 const emailOfInput = document.querySelector(".js-form-email");
 const passwordOfInput = document.querySelector(".js-form-password");
 const formElements = [nameOfInput, emailOfInput, passwordOfInput];
-
-const validationTerms = {
-    name: {
-        upperLimitOfText: 15
-    },
-    password: {
-        pattern: /^(?=.*?[a-z])(?=.*?[A-Z])(?=.*?\d)[a-zA-Z\d]{8,}$/
-    }
-}
-
-const validationOptions = {
-    name: {
-        isValid: () => {
-            return nameOfInput.value.length < validationTerms.name.upperLimitOfText;
-        },
-        errorMessage: "ユーザー名は15文字以内で入力してください",
-    },
-    email: {
-        isValid: () => {
-            return emailOfInput.validity.valid;
-        },
-        errorMessage: "メールアドレスの形式になっていません",
-    },
-    password: {
-        isValid: () => {
-            return validationTerms.password.pattern.test(passwordOfInput.value);
-        },
-        errorMessage: "8文字以上の大小の英数字を交ぜたものにしてください",
-    }
-};
 
 const addInvalidClass = target => target.classList.add("invalid");
 const removeInvalidClass = target => target.classList.remove("invalid");
