+++ lesson26/src/js/validation.js
@@ -1,0 +1,32 @@
+const nameOfInput = document.querySelector(".js-form-name");
+const emailOfInput = document.querySelector(".js-form-email");
+const passwordOfInput = document.querySelector(".js-form-password");
+const validationTerms = {
+    name: {
+        upperLimitOfText: 15
+    },
+    password: {
+        pattern: /^(?=.*?[a-z])(?=.*?[A-Z])(?=.*?\d)[a-zA-Z\d]{8,}$/
+    }
+}
+
+export const validationOptions = {
+    name: {
+        isValid: () => {
+            return nameOfInput.value.length < validationTerms.name.upperLimitOfText;
+        },
+        errorMessage: "ユーザー名は15文字以内で入力してください",
+    },
+    email: {
+        isValid: () => {
+            return emailOfInput.validity.valid;
+        },
+        errorMessage: "メールアドレスの形式になっていません",
+    },
+    password: {
+        isValid: () => {
+            return validationTerms.password.pattern.test(passwordOfInput.value);
+        },
+        errorMessage: "8文字以上の大小の英数字を交ぜたものにしてください",
+    }
+};
