+++ lesson26/src/register.html
@@ -6,7 +6,7 @@
         <meta http-equiv="X-UA-Compatible" content="ie=edge">
         <link rel="stylesheet" href="./css/reset.css">
         <link rel="stylesheet" href="./css/style.css">
-        <script src="./js/register.js" defer></script>
+        <script type="module" src="./js/register.js"></script>
         <title>Sign Up</title>
     </head>
     <body>
